raise exception if trying to start niet processes before the pid dir has been setup; bump version

+++ lib/easy/deployment/niet.rb
@@ -4,6 +4,10 @@
 Capistrano::Configuration.instance(:must_exist).load do
   set :niet_process_count, 2
 
+  def remote_file_exists?(full_path)
+    'true' ==  capture("if [ -e #{full_path} ]; then echo 'true'; fi").strip
+  end
+
   namespace :niet do
     desc "Setup Niet"
     task :setup, roles: :job do
@@ -12,14 +16,22 @@ Capistrano::Configuration.instance(:must_exist).load do
 
     desc "Starts the niet process monitor and its jobs"
     task :start, roles: :job do
-      niet_process_count.times do |i|
-        run "niet -p #{shared_path}/niet/jobs_worker_#{i}.pid -c #{current_path} bundle exec rake jobs:work RAILS_ENV=#{stage}"
+      if remote_file_exists?("#{shared_path}/niet")
+        niet_process_count.times do |i|
+          run "niet -p #{shared_path}/niet/jobs_worker_#{i}.pid -c #{current_path} bundle exec rake jobs:work RAILS_ENV=#{stage}"
+        end
+      else
+        raise StandardError, "shared niet directory doesn't exist! Please run `cap #{stage} niet:setup` first"
       end
     end
 
     desc "Restarts the processes running under niet"
     task :restart, roles: :job do
-      run "for job in #{shared_path}/niet/* ; do kill -TERM `cat $job`; done"
+      if remote_file_exists?("#{shared_path}/niet")
+        run "for job in #{shared_path}/niet/* ; do kill -TERM `cat $job`; done"
+      else
+        raise StandardError, "shared niet directory doesn't exist! Please run `cap #{stage} niet:setup` first"
+      end
     end
 
     desc "Stops the processes running under niet and the niet process monitor"
